api/requirements.txt
# FastAPI core
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.6.1

# File handling & async
aiofiles==23.2.1
python-dotenv==1.0.1

# Async task queue
celery[redis]==5.3.6
redis==5.0.3

# OCR & PDF processing
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0

# OpenAI
openai>=1.12.0

# Auth & Security
python-jose==3.3.0
bcrypt==4.1.2
email-validator==2.1.1

# MongoDB
pymongo==4.6.1

# ============================
# SYSTEM DEPENDENCIES (Required)
# ============================
# Tesseract OCR (with Tamil language pack)
#   Ubuntu: sudo apt install tesseract-ocr tesseract-ocr-tam
#   macOS:  brew install tesseract
#
# Poppler (required by pdf2image)
#   Ubuntu: sudo apt install poppler-utils
#   macOS:  brew install poppler
reportlab==4.0.9
